requirements.txt
# SageAlpha.ai v3.0 - 2025 Flask Stack
# Pinned versions for production stability

# ==================== Core Flask ====================
Flask==3.1.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-CORS==5.0.0
Flask-SocketIO==5.4.1
Flask-Limiter==3.8.0
Flask-WTF==1.2.2
Flask-JWT-Extended==4.7.1

# ==================== Database ====================
SQLAlchemy==2.0.36
Flask-Migrate==4.0.7
alembic==1.14.0
# PostgreSQL support (production)
psycopg2-binary==2.9.10

# ==================== Security ====================
Werkzeug>=3.1.0
bcrypt==4.2.1
python-dotenv==1.0.1
PyJWT==2.10.1

# ==================== Azure Services ====================
azure-storage-blob>=12.24.0
azure-search-documents==11.6.0
azure-identity==1.19.0

# ==================== OpenAI & AI ====================
openai>=1.58.0
httpx==0.28.1

# ==================== PDF Processing ====================
PyPDF2==3.0.1
playwright==1.49.1

# ==================== Data Processing ====================
numpy==2.2.0
pandas==2.2.3

# ==================== Web Server (Production) ====================
gunicorn==23.0.0
gevent==24.11.1
gevent-websocket==0.10.1
eventlet==0.37.0

# ==================== HTTP Client ====================
requests==2.32.3

# ==================== Background Tasks (Celery + Redis) ====================
celery[redis]==5.4.0
redis==5.2.1
kombu==5.4.2

# ==================== Utilities ====================
six==1.17.0
python-dateutil==2.9.0

# ==================== Development & Code Quality ====================
black==24.10.0
isort==5.13.2
flake8==7.1.1
mypy==1.13.0
pytest==8.3.4
pytest-cov==6.0.0
